refactor(common,lld,llm): adjust errors based on feedback + fix unfreezenotexpired logic
