+++ apps/ledger-live-desktop/static/i18n/en/app.json
@@ -5616,7 +5616,7 @@
       "title": "No assets to unfreeze",
       "description": "You do not have assets to unfreeze associated to Energy"
     },
-    "TronUnfreezeNotExpired": {
+    "TronLegacyUnfreezeNotExpired": {
       "title": "Unfreeze is not available yet",
       "description": "You need to wait 3 days after your last Freeze operation"
     },
@@ -5653,6 +5653,16 @@
     },
     "TronNotEnoughEnergy": {
       "title": "Energy is lower than necessary. You might pay up to 50 TRX in fees."
+    },
+    "TronNoUnfrozenResource": {
+      "title": "No unfrozen resources to withdraw"
+    },
+    "TronInvalidUnDelegateResourceAmount": {
+      "title": "Not enough delegated resources"
+    },
+    "TronUnfreezeNotExpired": {
+      "title": "Resources not unfrozen yet",
+      "description": "You need to wait until {{time}} to withdraw"
     },
     "PairingFailed": {
       "title": "Pairing unsuccessful",
