+++ apps/ledger-live-desktop/static/i18n/fr/app.json
@@ -5576,7 +5576,7 @@
       "title" : "Aucun actif à dégeler",
       "description" : "Vous n’avez pas d’actifs associés à l’Énergie à dégeler."
     },
-    "TronUnfreezeNotExpired" : {
+    "TronLegacyUnfreezeNotExpired" : {
       "title" : "Dégel indisponible pour le moment",
       "description" : "Merci de patienter 3 jours après votre dernière opération de gel."
     },
