+++ apps/ledger-live-desktop/static/i18n/ja/app.json
@@ -5576,7 +5576,7 @@
       "title" : "凍結解除する資産がありません。",
       "description" : "Energy(エネルギー)に連携している凍結解除アカウントがありません。"
     },
-    "TronUnfreezeNotExpired" : {
+    "TronLegacyUnfreezeNotExpired" : {
       "title" : "凍結解除はまだ利用できません。",
       "description" : "前回の凍結から3日後にご利用ください。"
     },
