+++ apps/ledger-live-desktop/static/i18n/ru/app.json
@@ -5576,7 +5576,7 @@
       "title" : "Нет активов для разморозки",
       "description" : "У вас нет активов для разморозки, работающих от Energy"
     },
-    "TronUnfreezeNotExpired" : {
+    "TronLegacyUnfreezeNotExpired" : {
       "title" : "Функция разморозки пока недоступна",
       "description" : "После заморозки активов должно пройти 3 дня."
     },
