+++ apps/ledger-live-desktop/static/i18n/zh/app.json
@@ -5576,7 +5576,7 @@
       "title" : "没有可解冻的资产",
       "description" : "您没有与能量相关联的资产可解冻"
     },
-    "TronUnfreezeNotExpired" : {
+    "TronLegacyUnfreezeNotExpired" : {
       "title" : "暂时无法解冻",
       "description" : "上次冻结操作后，需要等待 3 天"
     },
