+++ apps/ledger-live-mobile/src/locales/ko/common.json
@@ -556,7 +556,7 @@
       "title" : "동결 해제할 자산 없음",
       "description" : "동결 해제할 에너지 자산이 없습니다."
     },
-    "TronUnfreezeNotExpired" : {
+    "TronLegacyUnfreezeNotExpired" : {
       "title" : "아직 동결 해제할 수 없습니다",
       "description" : "마지막 동결 후 3일을 기다리세요."
     },
