+++ apps/ledger-live-mobile/src/locales/pt-BR/common.json
@@ -556,7 +556,7 @@
       "title" : "Sem ativos para descongelar",
       "description" : "Você não tem ativos de Energia para descongelar."
     },
-    "TronUnfreezeNotExpired" : {
+    "TronLegacyUnfreezeNotExpired" : {
       "title" : "Descongelamento ainda indisponível",
       "description" : "Aguarde 3 dias após sua última operação de congelamento."
     },
